backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["pony <pony8674312@gmail.com>"]

[tool.poetry.dependencies]
python = ">=3.10, <3.11"
fastapi = "^0.75.0"
uvicorn = {extras = ["standard"], version = "^0.17.6"}
psycopg2-binary = "^2.9.3"
FastAPI-SQLAlchemy = "^0.2.1"
plotly = "^5.10.0"
Levenshtein = "^0.21.1"
pandas = "^2.1.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.1.13"]
build-backend = "poetry.core.masonry.api"
